refactor(scan): extract `ballot-interpreter` package
This allows the `rave` repo to remove the ballot interpreter stuff but keep the `backend` library.
