+++ apps/central-scan/backend/tsconfig.json
@@ -20,6 +20,7 @@
     { "path": "../../../libs/api/tsconfig.build.json" },
     { "path": "../../../libs/auth/tsconfig.build.json" },
     { "path": "../../../libs/ballot-encoder/tsconfig.build.json" },
+    { "path": "../../../libs/ballot-interpreter/tsconfig.build.json" },
     { "path": "../../../libs/basics/tsconfig.build.json" },
     { "path": "../../../libs/converter-nh-accuvote/tsconfig.build.json" },
     { "path": "../../../libs/backend/tsconfig.build.json" },
