+++ apps/scan/backend/src/interpret.ts
@@ -10,7 +10,7 @@ import {
 } from '@votingworks/types';
 import { time } from '@votingworks/utils';
 import { err, ok, Optional, Result } from '@votingworks/basics';
-import { interpretSheetAndSaveImages } from '@votingworks/backend';
+import { interpretSheetAndSaveImages } from '@votingworks/ballot-interpreter';
 import { rootDebug } from './util/debug';
 import { SheetInterpretation } from './types';
