+++ libs/backend/src/scan/index.ts
@@ -3,6 +3,4 @@ export * from './cast_vote_records/build_cast_vote_record';
 export * from './cast_vote_records/build_report_metadata';
 export * from './cast_vote_records/export';
 export * from './cast_vote_records/canonicalize';
-export * from './interpretation/interpret';
-export * from './interpretation/interpret_and_save_files';
 export { VX_MACHINE_ID } from './globals';
