+++ libs/ballot-interpreter/.eslintignore
@@ -1,0 +1,3 @@
+build
+coverage
+jest.config.js
