+++ libs/ballot-interpreter/.eslintrc.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["plugin:vx/recommended"]
+}
