+++ libs/ballot-interpreter/README.md
@@ -1,0 +1,7 @@
+# @votingworks/ballot-interpreter
+
+Wrapper package for our various ballot interpreters.
+
+## License
+
+AGPLv3
