+++ libs/ballot-interpreter/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@votingworks/backend",
+  "name": "@votingworks/ballot-interpreter",
   "version": "1.0.0",
   "private": true,
-  "description": "Data manipulation utilities for NodeJS",
-  "license": "GPL-3.0",
+  "description": "Interpreter for scanned ballot images",
+  "license": "AGPL-3.0",
   "author": "VotingWorks Eng <eng@voting.works>",
   "main": "build/index.js",
   "types": "build/index.d.js",
@@ -17,9 +17,9 @@
     "lint:fix": "pnpm type-check && eslint . --fix",
     "pre-commit": "lint-staged",
     "test": "is-ci test:ci test:watch",
-    "test:ci": "TZ=UTC pnpm build && pnpm test:coverage --reporters=default --reporters=jest-junit --maxWorkers=7",
-    "test:coverage": "TZ=UTC jest --coverage",
-    "test:watch": "TZ=UTC jest --watch",
+    "test:ci": "pnpm build && pnpm test:coverage --reporters=default --reporters=jest-junit --maxWorkers=7",
+    "test:coverage": "jest --coverage",
+    "test:watch": "jest --watch",
     "type-check": "tsc --build"
   },
   "lint-staged": {
@@ -34,38 +34,23 @@
     ]
   },
   "dependencies": {
-    "@votingworks/auth": "workspace:*",
+    "@votingworks/ballot-interpreter-nh": "workspace:*",
+    "@votingworks/ballot-interpreter-vx": "workspace:*",
     "@votingworks/basics": "workspace:*",
     "@votingworks/fixtures": "workspace:*",
     "@votingworks/image-utils": "workspace:*",
-    "@votingworks/logging": "workspace:*",
     "@votingworks/types": "workspace:*",
     "@votingworks/utils": "workspace:*",
-    "buffer": "^6.0.3",
-    "debug": "^4.3.2",
-    "fs-extra": "^11.1.1",
-    "jszip": "^3.9.1",
-    "micromatch": "^4.0.5",
-    "readline": "^1.3.0",
-    "stream-chain": "^2.2.5",
-    "stream-chopper": "^3.0.1",
-    "stream-json": "^1.7.5",
-    "tmp": "^0.2.1",
-    "zod": "3.14.4"
+    "debug": "^4.3.2"
   },
   "devDependencies": {
     "@types/debug": "^4.1.7",
     "@types/fs-extra": "^11.0.1",
     "@types/jest": "^27.4.1",
-    "@types/micromatch": "^4.0.2",
     "@types/node": "16.18.23",
-    "@types/stream-chain": "^2.0.1",
-    "@types/stream-chopper": "workspace:*",
-    "@types/stream-json": "^1.7.3",
     "@types/tmp": "^0.2.3",
     "@typescript-eslint/eslint-plugin": "5.37.0",
     "@typescript-eslint/parser": "5.37.0",
-    "@votingworks/test-utils": "workspace:*",
     "eslint": "8.23.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-node": "^0.3.6",
@@ -73,16 +58,16 @@
     "eslint-plugin-jest": "^26.1.5",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-vx": "workspace:*",
-    "fast-check": "2.23.2",
+    "fs-extra": "^11.1.1",
     "is-ci-cli": "^2.2.0",
     "jest": "^27.3.1",
     "jest-date-mock": "^1.0.8",
     "jest-junit": "^14.0.1",
     "jest-watch-typeahead": "^0.6.4",
     "lint-staged": "^11.0.0",
-    "memory-streams": "^0.1.3",
     "prettier": "2.6.2",
     "sort-package-json": "^1.50.0",
+    "tmp": "^0.2.1",
     "ts-jest": "^27.0.7",
     "typescript": "4.6.3"
   },
