+++ libs/ballot-interpreter/src/index.ts
@@ -1,0 +1,3 @@
+/* istanbul ignore file */
+export * from './interpret';
+export * from './interpret_and_save_files';
