+++ libs/ballot-interpreter/src/interpret.test.ts
