+++ libs/ballot-interpreter/src/interpret_and_save_files.test.ts
@@ -3,7 +3,7 @@ import { SheetOf } from '@votingworks/types';
 import { ALL_PRECINCTS_SELECTION } from '@votingworks/utils';
 import { loadImageData } from '@votingworks/image-utils';
 import { interpretSheetAndSaveImages } from './interpret_and_save_files';
-import { tmpDir } from '../../../test/helpers/tmp';
+import { tmpDir } from '../test/helpers/tmp';
 
 test('interprets ballot images and saves images for storage', async () => {
   const fixtures = electionFamousNames2021Fixtures;
