+++ libs/ballot-interpreter/src/legacy_adapter.ts
