+++ libs/ballot-interpreter/src/save_images.test.ts
@@ -5,7 +5,7 @@ import {
   writeImageData,
 } from '@votingworks/image-utils';
 import { join } from 'path';
-import { tmpDir } from '../../../test/helpers/tmp';
+import { tmpDir } from '../test/helpers/tmp';
 import { saveSheetImage } from './save_images';
 
 test.each([
