+++ libs/ballot-interpreter/src/save_images.ts
